requirements.txt
opencv-python>=4.8.0
deepface>=0.0.79
tensorflow>=2.13.0
tf-keras>=2.20.0
pillow>=10.0.0
numpy>=1.24.0
